+++ apps/api/src/app/endpoints/public/public.controller.ts
@@ -1,0 +1,129 @@
+import { AccessService } from '@ghostfolio/api/app/access/access.service';
+import { PortfolioService } from '@ghostfolio/api/app/portfolio/portfolio.service';
+import { UserService } from '@ghostfolio/api/app/user/user.service';
+import { TransformDataSourceInResponseInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.interceptor';
+import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
+import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
+import { DEFAULT_CURRENCY } from '@ghostfolio/common/config';
+import { PublicPortfolioResponse } from '@ghostfolio/common/interfaces';
+import type { RequestWithUser } from '@ghostfolio/common/types';
+
+import {
+  Controller,
+  Get,
+  HttpException,
+  Inject,
+  Param,
+  UseInterceptors
+} from '@nestjs/common';
+import { REQUEST } from '@nestjs/core';
+import { StatusCodes, getReasonPhrase } from 'http-status-codes';
+
+@Controller('public')
+export class PublicController {
+  public constructor(
+    private readonly accessService: AccessService,
+    private readonly configurationService: ConfigurationService,
+    private readonly exchangeRateDataService: ExchangeRateDataService,
+    private readonly portfolioService: PortfolioService,
+    @Inject(REQUEST) private readonly request: RequestWithUser,
+    private readonly userService: UserService
+  ) {}
+
+  @Get(':accessId/portfolio')
+  @UseInterceptors(TransformDataSourceInResponseInterceptor)
+  public async getPublicPortfolio(
+    @Param('accessId') accessId
+  ): Promise<PublicPortfolioResponse> {
+    const access = await this.accessService.access({ id: accessId });
+
+    if (!access) {
+      throw new HttpException(
+        getReasonPhrase(StatusCodes.NOT_FOUND),
+        StatusCodes.NOT_FOUND
+      );
+    }
+
+    let hasDetails = true;
+
+    const user = await this.userService.user({
+      id: access.userId
+    });
+
+    if (this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION')) {
+      hasDetails = user.subscription.type === 'Premium';
+    }
+
+    const [
+      { holdings },
+      { performance: performance1d },
+      { performance: performanceMax },
+      { performance: performanceYtd }
+    ] = await Promise.all([
+      this.portfolioService.getDetails({
+        filters: [{ id: 'EQUITY', type: 'ASSET_CLASS' }],
+        impersonationId: access.userId,
+        userId: user.id,
+        withMarkets: true
+      }),
+      ...['1d', 'max', 'ytd'].map((dateRange) => {
+        return this.portfolioService.getPerformance({
+          dateRange,
+          impersonationId: undefined,
+          userId: user.id
+        });
+      })
+    ]);
+
+    const publicPortfolioResponse: PublicPortfolioResponse = {
+      hasDetails,
+      alias: access.alias,
+      holdings: {},
+      performance: {
+        '1d': {
+          relativeChange:
+            performance1d.netPerformancePercentageWithCurrencyEffect
+        },
+        max: {
+          relativeChange:
+            performanceMax.netPerformancePercentageWithCurrencyEffect
+        },
+        ytd: {
+          relativeChange:
+            performanceYtd.netPerformancePercentageWithCurrencyEffect
+        }
+      }
+    };
+
+    const totalValue = Object.values(holdings)
+      .map(({ currency, marketPrice, quantity }) => {
+        return this.exchangeRateDataService.toCurrency(
+          quantity * marketPrice,
+          currency,
+          this.request.user?.Settings?.settings.baseCurrency ?? DEFAULT_CURRENCY
+        );
+      })
+      .reduce((a, b) => a + b, 0);
+
+    for (const [symbol, portfolioPosition] of Object.entries(holdings)) {
+      publicPortfolioResponse.holdings[symbol] = {
+        allocationInPercentage:
+          portfolioPosition.valueInBaseCurrency / totalValue,
+        countries: hasDetails ? portfolioPosition.countries : [],
+        currency: hasDetails ? portfolioPosition.currency : undefined,
+        dataSource: portfolioPosition.dataSource,
+        dateOfFirstActivity: portfolioPosition.dateOfFirstActivity,
+        markets: hasDetails ? portfolioPosition.markets : undefined,
+        name: portfolioPosition.name,
+        netPerformancePercentWithCurrencyEffect:
+          portfolioPosition.netPerformancePercentWithCurrencyEffect,
+        sectors: hasDetails ? portfolioPosition.sectors : [],
+        symbol: portfolioPosition.symbol,
+        url: portfolioPosition.url,
+        valueInPercentage: portfolioPosition.valueInBaseCurrency / totalValue
+      };
+    }
+
+    return publicPortfolioResponse;
+  }
+}
