+++ apps/api/src/app/endpoints/public/public.module.ts
@@ -1,0 +1,49 @@
+import { AccessModule } from '@ghostfolio/api/app/access/access.module';
+import { AccountBalanceService } from '@ghostfolio/api/app/account-balance/account-balance.service';
+import { AccountService } from '@ghostfolio/api/app/account/account.service';
+import { OrderModule } from '@ghostfolio/api/app/order/order.module';
+import { PortfolioCalculatorFactory } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator.factory';
+import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
+import { PortfolioService } from '@ghostfolio/api/app/portfolio/portfolio.service';
+import { RulesService } from '@ghostfolio/api/app/portfolio/rules.service';
+import { RedisCacheModule } from '@ghostfolio/api/app/redis-cache/redis-cache.module';
+import { UserModule } from '@ghostfolio/api/app/user/user.module';
+import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
+import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
+import { ExchangeRateDataModule } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.module';
+import { ImpersonationModule } from '@ghostfolio/api/services/impersonation/impersonation.module';
+import { MarketDataModule } from '@ghostfolio/api/services/market-data/market-data.module';
+import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
+import { PortfolioSnapshotQueueModule } from '@ghostfolio/api/services/queues/portfolio-snapshot/portfolio-snapshot.module';
+import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/symbol-profile.module';
+
+import { Module } from '@nestjs/common';
+
+import { PublicController } from './public.controller';
+
+@Module({
+  controllers: [PublicController],
+  imports: [
+    AccessModule,
+    DataProviderModule,
+    ExchangeRateDataModule,
+    ImpersonationModule,
+    MarketDataModule,
+    OrderModule,
+    PortfolioSnapshotQueueModule,
+    PrismaModule,
+    RedisCacheModule,
+    SymbolProfileModule,
+    TransformDataSourceInRequestModule,
+    UserModule
+  ],
+  providers: [
+    AccountBalanceService,
+    AccountService,
+    CurrentRateService,
+    PortfolioCalculatorFactory,
+    PortfolioService,
+    RulesService
+  ]
+})
+export class PublicModule {}
