+++ apps/client/src/app/components/access-table/access-table.component.html
@@ -44,9 +44,9 @@
           @if (user?.settings?.isExperimentalFeatures) {
             <div>
               <code
-                >GET {{ baseUrl }}/api/v1/portfolio/public/{{
+                >GET {{ baseUrl }}/api/v1/public/{{
                   element.id
-                }}</code
+                }}/portfolio</code
               >
             </div>
           }
