+++ apps/client/src/app/pages/public/public-page.component.ts
@@ -3,7 +3,7 @@ import { UNKNOWN_KEY } from '@ghostfolio/common/config';
 import { prettifySymbol } from '@ghostfolio/common/helper';
 import {
   PortfolioPosition,
-  PortfolioPublicResponse
+  PublicPortfolioResponse
 } from '@ghostfolio/common/interfaces';
 import { Market } from '@ghostfolio/common/types';
 
@@ -29,16 +29,16 @@ export class PublicPageComponent implements OnInit {
     [code: string]: { name: string; value: number };
   };
   public deviceType: string;
-  public holdings: PortfolioPublicResponse['holdings'][string][];
+  public holdings: PublicPortfolioResponse['holdings'][string][];
   public markets: {
     [key in Market]: { name: string; value: number };
   };
-  public portfolioPublicDetails: PortfolioPublicResponse;
   public positions: {
     [symbol: string]: Pick<PortfolioPosition, 'currency' | 'name'> & {
       value: number;
     };
   };
+  public publicPortfolioDetails: PublicPortfolioResponse;
   public sectors: {
     [name: string]: { name: string; value: number };
   };
@@ -47,7 +47,7 @@ export class PublicPageComponent implements OnInit {
   };
   public UNKNOWN_KEY = UNKNOWN_KEY;
 
-  private id: string;
+  private accessId: string;
   private unsubscribeSubject = new Subject<void>();
 
   public constructor(
@@ -58,7 +58,7 @@ export class PublicPageComponent implements OnInit {
     private router: Router
   ) {
     this.activatedRoute.params.subscribe((params) => {
-      this.id = params['id'];
+      this.accessId = params['id'];
     });
   }
 
@@ -66,7 +66,7 @@ export class PublicPageComponent implements OnInit {
     this.deviceType = this.deviceService.getDeviceInfo().deviceType;
 
     this.dataService
-      .fetchPortfolioPublic(this.id)
+      .fetchPublicPortfolio(this.accessId)
       .pipe(
         takeUntil(this.unsubscribeSubject),
         catchError((error) => {
@@ -79,7 +79,7 @@ export class PublicPageComponent implements OnInit {
         })
       )
       .subscribe((portfolioPublicDetails) => {
-        this.portfolioPublicDetails = portfolioPublicDetails;
+        this.publicPortfolioDetails = portfolioPublicDetails;
 
         this.initializeAnalysisData();
 
@@ -135,7 +135,7 @@ export class PublicPageComponent implements OnInit {
     };
 
     for (const [symbol, position] of Object.entries(
-      this.portfolioPublicDetails.holdings
+      this.publicPortfolioDetails.holdings
     )) {
       this.holdings.push(position);
 
@@ -164,7 +164,7 @@ export class PublicPageComponent implements OnInit {
               name: continent,
               value:
                 weight *
-                this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency
+                this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency
             };
           }
 
@@ -175,19 +175,19 @@ export class PublicPageComponent implements OnInit {
               name,
               value:
                 weight *
-                this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency
+                this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency
             };
           }
         }
       } else {
         this.continents[UNKNOWN_KEY].value +=
-          this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency;
+          this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency;
 
         this.countries[UNKNOWN_KEY].value +=
-          this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency;
+          this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency;
 
         this.markets[UNKNOWN_KEY].value +=
-          this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency;
+          this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency;
       }
 
       if (position.sectors.length > 0) {
@@ -201,13 +201,13 @@ export class PublicPageComponent implements OnInit {
               name,
               value:
                 weight *
-                this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency
+                this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency
             };
           }
         }
       } else {
         this.sectors[UNKNOWN_KEY].value +=
-          this.portfolioPublicDetails.holdings[symbol].valueInBaseCurrency;
+          this.publicPortfolioDetails.holdings[symbol].valueInBaseCurrency;
       }
 
       this.symbols[prettifySymbol(symbol)] = {
