+++ libs/common/src/lib/interfaces/index.ts
@@ -43,7 +43,7 @@ import type { ImportResponse } from './responses/import-response.interface';
 import type { OAuthResponse } from './responses/oauth-response.interface';
 import type { PortfolioHoldingsResponse } from './responses/portfolio-holdings-response.interface';
 import type { PortfolioPerformanceResponse } from './responses/portfolio-performance-response.interface';
-import type { PortfolioPublicResponse } from './responses/portfolio-public-response.interface';
+import type { PublicPortfolioResponse } from './responses/public-portfolio-response.interface';
 import type { ScraperConfiguration } from './scraper-configuration.interface';
 import type { Statistics } from './statistics.interface';
 import type { Subscription } from './subscription.interface';
@@ -91,12 +91,12 @@ export {
   PortfolioPerformance,
   PortfolioPerformanceResponse,
   PortfolioPosition,
-  PortfolioPublicResponse,
   PortfolioReport,
   PortfolioReportRule,
   PortfolioSummary,
   Position,
   Product,
+  PublicPortfolioResponse,
   ResponseError,
   ScraperConfiguration,
   Statistics,
